+++ scripts/consts/GmmZoneAudiences.js
@@ -1,0 +1,5 @@
+export const GMM_ZONE_AUDIENCES = [
+	"any",
+	"ally",
+	"enemy"
+];