+++ scripts/consts/GmmZonePayloads.js
@@ -1,0 +1,5 @@
+export const GMM_ZONE_PAYLOADS = [
+	"damage",
+	"attack",
+	"effects"
+];